+++ src/features/settings/model/local-storage.ts
@@ -7,7 +7,7 @@
 
 import { DARK_THEME, GsLang, GsLangUser, GsTheme, LANG_SYSTEM } from '@gridsuite/commons-ui';
 import { getComputedLanguage } from '../../../shared/lib/language';
-import { APP_NAME } from '../../../shared/lib/config-params';
+import { APP_NAME } from '../../../app/config/app';
 
 const LOCAL_STORAGE_THEME_KEY = (APP_NAME + '_THEME').toUpperCase();
 const LOCAL_STORAGE_LANGUAGE_KEY = (APP_NAME + '_LANGUAGE').toUpperCase();
