+++ src/features/topbar/useAppTopBar.ts
@@ -12,8 +12,8 @@ import { useNavigate } from 'react-router-dom';
 import type { AppTopBarProps } from '../../app/layout/AppTopBar';
 import { AppDispatch } from '../../app/store';
 import { selectTheme } from '../settings/model/selectors';
-import { PARAM_LANGUAGE, PARAM_THEME } from '../../shared/lib/config-params';
-import { getServersInfos } from '../../shared/lib/study';
+import { PARAM_LANGUAGE, PARAM_THEME } from '../../shared/config/parameters';
+import { getServersInfos } from './api/study';
 import { useParameterState } from './useParameterState';
 import { useTopBarApps } from './useTopBarApps';
 import { useTopBarVersion } from './useTopBarVersion';
