+++ src/features/topbar/useParameterState.ts
@@ -9,7 +9,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { RootState } from '../../app/store';
-import { updateConfigParameter } from '../../shared/lib/rest-api';
+import { updateConfigParameter } from '../../shared/api/config';
 import { AppStateKey, SettingsState } from '../settings/model/types';
 
 export function useParameterState<K extends AppStateKey>(
