+++ src/shared/config/parameters.ts
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+export const PARAM_THEME = 'theme';
+export const PARAM_LANGUAGE = 'language';
+
+export type AppConfigParameter = typeof PARAM_THEME | typeof PARAM_LANGUAGE | string;
